% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{addWaterDay}
\alias{addWaterDay}
\title{Add water day}
\usage{
addWaterDay(df, date_colname, wy_colname)
}
\arguments{
\item{df}{data.frame: dataframe to add waterday. Must have waterYear and
date column}

\item{date_colname}{name of the df column of class "Date"}

\item{wy_colname}{name of the df column of class "numeric". It is
suggested to use dataRetrieval;addWaterYear to create this column.}
}
\value{
data.frame: original dataframe with waterDay column added
}
\description{
Create a trial child directory for a project and generate parameter files.
Note that for large morris or mmlhs trials, it may take a long time for the
function to complete.
}
\examples{

df_w_waterday <- addWaterDay(df,
  date_colname = Date,
  wy_colname = waterYear
)

}
