% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{calc_gof}
\alias{calc_gof}
\title{Calculate Goodness of Fit for pws}
\usage{
calc_gof(
  dir_root,
  project_name,
  trial_number,
  start_year,
  end_year,
  hru_out_names = c("pkwater_equiv", "snowcov_area", "hru_actet", "soil_rechr"),
  seg_out_names = "seg_outflow",
  default,
  aso = FALSE,
  smapl4 = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{dir_root}{character: parent working directory path.}

\item{project_name}{character: project name.}

\item{trial_number}{integer: trial number, leading zeros not needed.}

\item{start_year}{integer; first water year analyzed.}

\item{end_year}{integer: last water year analyzed.}

\item{hru_out_names}{character: vector of pws hru output file names (without
file extension e.g. "pkwater_equiv").}

\item{seg_out_names}{character: vector of pws segment output file names
(without file extension e.g "seg_outflow").}

\item{default}{logical: set to TRUE when analyzing the single default run.}

\item{aso}{logical: set to TRUE when aso is to be included in analysis.}

\item{smapl4}{logical: set to TRUE when smap is to be included in analysis.}

\item{overwrite}{logical: set to TRUE when existing analysis results are
to be overwritten.}
}
\value{
list: nested list of goodness of fit results
}
\description{
This function reads pws outputs and compares them to observations.
}
