% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pws.R
\name{.check_existing_runs}
\alias{.check_existing_runs}
\title{Function to check for existing folders and adjust runs to avoid overwriting}
\usage{
.check_existing_runs(runs, dir_dynamic_output)
}
\description{
Function to check for existing folders and adjust runs to avoid overwriting
}
