% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morris.R
\name{.compute_measures}
\alias{.compute_measures}
\title{Compute mu, mu_star, and sigma for EEs}
\usage{
.compute_measures(ee_matrix)
}
\description{
Based on documentation in the sensitivity package
https://cran.r-project.org/web/packages/sensitivity/sensitivity.pdf
page 31
}
