% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_create_attributes}
\alias{.ig_create_attributes}
\title{Retrieve pws parameter attributes}
\usage{
.ig_create_attributes(param_names)
}
\arguments{
\item{param_names}{character vector: names of pws calibration parameters}
}
\value{
data.frame: where row 1 and row 2 are min, max values respectively
}
\description{
This function takes your specified working directory and creates a child
directory specific to the project name and relevant input/output directories
}
