% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_dims_read}
\alias{.ig_dims_read}
\title{Read .param file dimensions}
\usage{
.ig_dims_read(file_path)
}
\arguments{
\item{file_path}{character: string containing full path of target param file}
}
\value{
dimensions list: a list containing the dimensions section
}
\description{
This function parses the dimensions section of a legacy PRMS .param file
}
