% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_dims_write}
\alias{.ig_dims_write}
\title{Write .param file dimensions}
\usage{
.ig_dims_write(file_path, dimensions, run)
}
\arguments{
\item{file_path}{character: string containing full path of target param file}

\item{dimensions}{list: contains the dimensions from .ig_read_dims}

\item{run}{integer: declares run number when run inside .ig_sample functions}
}
\value{
dimensions list: a list containing the dimensions section
}
\description{
This function writes the dimensions section of a legacy PRMS .param file.
This section is not modified during parameter file generation.
}
