% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_normalize_minmax}
\alias{.ig_normalize_minmax}
\title{Normalize absolute parameter values}
\usage{
.ig_normalize_minmax(x_a, max_x_a, min_x_a)
}
\arguments{
\item{x_a}{numeric: vector of absolute param values}

\item{max_x_a}{numeric: the max value of x_a}

\item{min_x_a}{numeric: the min value of x_a}
}
\value{
x_n numeric: a vector of param values normalized to 0 and 1
}
\description{
This function normalizes parameter values between 0 and 1
}
