% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_normalize_reverse}
\alias{.ig_normalize_reverse}
\title{Reverse 0 to 1 normalization}
\usage{
.ig_normalize_reverse(x_n, max_x_a, min_x_a)
}
\arguments{
\item{x_n}{numeric: vector of normalized param values}

\item{max_x_a}{numeric: the max value of x_a}

\item{min_x_a}{numeric: the min value of x_a}
}
\value{
x_a numeric: a vector of actual param values
}
\description{
This function reverses 0 to 1 normalization to absolute magnitudes
}
