% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_params_read}
\alias{.ig_params_read}
\title{Read .param file parameters}
\usage{
.ig_params_read(file_path, dimensions)
}
\arguments{
\item{file_path}{character: string containing full path of target param file}

\item{dimensions}{list: contains info on dimension sizes from .ig_read_dims}
}
\value{
parameters list: a list containing the parameters section
}
\description{
This function parses the parameters section of a legacy PRMS .param file.
The format is standardized, based on USGS documentation
https://pubs.usgs.gov/tm/6b7/pdf/tm6-b7.pdf page 129
}
