% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_params_write}
\alias{.ig_params_write}
\title{Write .param file parameters}
\usage{
.ig_params_write(parameters, file_path)
}
\arguments{
\item{parameters}{list: contains the parameters from .ig_read_params}

\item{file_path}{character: string containing full path of target param file}
}
\value{
parameters list: a list containing the parameters section
}
\description{
This function writes the parameters section of a legacy PRMS .param file.
}
