% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_perturb_params}
\alias{.ig_perturb_params}
\title{Perturb parameter values}
\usage{
.ig_perturb_params(
  sampling_matrix,
  dimensions,
  parameters,
  param_attributes,
  dir_dynamic_input
)
}
\arguments{
\item{sampling_matrix}{numeric: 2D matrix of normalized parameters.
The dimensions are nruns (rows) by nparams (cols).}

\item{dimensions}{list: a list containing the dimensions section}

\item{parameters}{list: a list containing the parameters section}

\item{param_attributes}{data.frame: contains columns named min, max}

\item{dir_dynamic_input}{character: path to where the param file is written
values respectively. Columns must match that of the sampling_matrix.}
}
\value{
NULL: written parameter file
}
\description{
This function creates new parameter values based on sampling values
produced from morris or mmlhs functions.
}
