% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_sample_mmlhs}
\alias{.ig_sample_mmlhs}
\title{Create mmlhs parameter files}
\usage{
.ig_sample_mmlhs(nruns, param_attributes, dir_default_input, dir_dynamic_input)
}
\arguments{
\item{nruns}{integer: Number of desired runs (should be 1000 * param number)}

\item{param_attributes}{data.frame: contains columns named min, max}

\item{dir_dynamic_input}{character: path where the param files are written}

\item{dir_defualt_input}{character: path where the original param file exists}
}
\value{
list_mmlhs list: Param values, sample matrix,
}
\description{
This function create perturbed parameter files via the Maximin Latin
Hypercube Sampling method
}
