% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_generation.R
\name{.ig_sample_morris}
\alias{.ig_sample_morris}
\title{Create morris parameter files}
\usage{
.ig_sample_morris(r, param_attributes, dir_default_input, dir_dynamic_input)
}
\arguments{
\item{r}{numeric: 2D matrix of normalized parameters.}

\item{param_attributes}{data.frame: contains columns named min, max}

\item{dir_dynamic_input}{character: path where the param files are written}

\item{dir_defualt_input}{character: path where the original param file exists}
}
\value{
list_morris list: Sampling design, param values, sample matrix,
}
\description{
This function create perturbed parameter files via the Morris Elementary
Effects method
}
