% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morris.R
\name{.load_json_if_exists}
\alias{.load_json_if_exists}
\title{Function to load JSON files conditionally based on variables and metrics}
\usage{
.load_json_if_exists(var, metric, dir_dynamic_output)
}
\description{
Function to load JSON files conditionally based on variables and metrics
}
