% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morris.R
\name{.logis_offset_curv}
\alias{.logis_offset_curv}
\title{Logistic function signed curvature}
\usage{
.logis_offset_curv(x, lo, aa, k, xo)
}
\arguments{
\item{x}{integer: param rank}

\item{lo}{numeric: min y value (eta* value)}

\item{aa}{numeric: max(y) - min(y)}

\item{k}{numeric: sigmoid steepness (calibrate from 1)}

\item{x0}{numeric: sigmoid midpoint (calibrate from 0.8)}
}
\value{
numeric: signed curvature of the sigmoid
}
\description{
See appendix of Cuntz et al., 2015
https://agupubs.onlinelibrary.wiley.com/doi/epdf/10.1002/2015WR016907
"Computationally inexpensive identification of noninformative model
parameters by sequential screening"
}
