% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.read_json}
\alias{.read_json}
\title{Helper to read different output list types, wrapped in 'read_output'}
\usage{
.read_json(list_type, var, metric, dir_dynamic_output)
}
\description{
Helper to read different output list types, wrapped in 'read_output'
}
