% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morris.R
\name{.recursive_apply}
\alias{.recursive_apply}
\title{Recursive function to apply EE and optionally bootstrap}
\usage{
.recursive_apply(lst, n_boot, sd_level, morris_design)
}
\description{
Recursive function to apply EE and optionally bootstrap
}
