% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pws.R
\name{.run_parallel}
\alias{.run_parallel}
\title{Helper to setup a parallel plan, wrapper for .run_simulation}
\usage{
.run_parallel(
  dir_root,
  project_name,
  trial_number,
  runs = NULL,
  pwsenvname = "pws",
  out_var_names = c("seg_outflow", "pkwater_equiv", "snowcov_area", "soil_rechr",
    "hru_actet"),
  performance = "medium"
)
}
\description{
Helper to setup a parallel plan, wrapper for .run_simulation
}
