% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pws.R
\name{.set_workers}
\alias{.set_workers}
\title{Function to determine worker count based on 'performance' level}
\usage{
.set_workers(performance)
}
\description{
Function to determine worker count based on 'performance' level
}
