% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morris.R
\name{.split_to_vectors}
\alias{.split_to_vectors}
\title{Helper function to split lists into vectors of length nruns}
\usage{
.split_to_vectors(lst)
}
\description{
For 3D arrays, the dims are nruns, nyears, nhru
}
