% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pws.R
\name{.temp_to_output}
\alias{.temp_to_output}
\title{Function to copy model output in temporary directory to model directory}
\usage{
.temp_to_output(out_var_names, temp_working_dir, model_dir)
}
\arguments{
\item{temp_working_dir}{character: created in .run_simulation}

\item{model_dir}{character: full file path of the .nc output file created
in .run_simulation}
}
\description{
Function to copy model output in temporary directory to model directory
}
