% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pws.R
\name{.truncate_runs}
\alias{.truncate_runs}
\title{Function to truncate runs to prevent exceeding nruns}
\usage{
.truncate_runs(runs, nruns)
}
\description{
Function to truncate runs to prevent exceeding nruns
}
