% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lhs.R
\name{find_behav}
\alias{find_behav}
\title{Find behavioral simulations}
\usage{
find_behav(
  df_gof,
  gof_thresh,
  cost,
  alt_thresh = NULL,
  alt_cols = NULL,
  diagnostic_hist = TRUE
)
}
\arguments{
\item{df_gof}{data.frame: A dataframe where the columns are observation
types, the rows are simulations, and values are GOF statistics.}

\item{gof_thresh}{numeric: A single value that you want to set as the
behavioral threshold.}

\item{cost}{logical: set to TRUE if using a cost function that should be
minimized (e.g. NRMSE, RMSE, PBIAS). Set to FALSE if using a likelihood
function that should be maximized (e.g. NSE, KGE).}

\item{alt_thresh}{numeric: An optional value if there are certain
observations that should follow and alternative threshold.}

\item{diagnostic_hist}{logical: Set to TRUE to view a histogram of
goodness of fit metrics in 'df_gof'.}

\item{alt_colnames}{character: A vector specifying the column names in
'df_gof' where the 'alt_thresh' should be applied.}
}
\value{
list: a nested list containing the indices and actual values of
behavioral simulations.
}
\description{
Filter the goodness of fit results via user specified thresholds
}
