% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lhs.R
\name{find_behav_intersect}
\alias{find_behav_intersect}
\title{Find behavioral simulation intersection}
\usage{
find_behav_intersect(df_gof, criteria_names, idx_behav, q_col = 1)
}
\arguments{
\item{df_gof}{data.frame: A dataframe where the columns are observation
types, the rows are simulations, and values are GOF statistics.}

\item{criteria_names}{list: a list of the criteria, where each element of the
list contains a set of column names in 'df_gof' to intersect.}

\item{idx_behav}{data.frame: one of the outputs from find_beahv. It contains
vectors of the simulation indices for behavioral simulations.}

\item{q_col}{integer: The column index in 'df_gof' which is used to return
a dataframe of the behavioral goodness of fit for that column. Generally
expects that Q is of interest.}
}
\value{
list: contains a dataframe if simulation indexes the meet the
intersection crtieria, as well as a dataframe of the GOF that correspond to
the selected 'q_col'.
}
\description{
Filter the goodness of fit results to intersections of multiple observations.
}
\details{
This should be run after find_behav and defining the multi-objective
criteria in a list.
}
