% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morris.R
\name{fit_logis}
\alias{fit_logis}
\title{Plot logistic function and find statistical errors in parameter sensitivity}
\usage{
fit_logis(
  df_mean,
  df_lower,
  df_upper,
  ncol_plot,
  dir_plot,
  start_xo = 0.8,
  start_k = 1,
  plot_filename = "plot_sensi_logis.png"
)
}
\arguments{
\item{df_mean}{data.frame: bootstrap mean eta* valeus where column names
correspond to the observation type and nrows in the the number of parameters.
The rows should be named by the parameters for more tangible results.}

\item{df_lower}{data.frame: bootstrap lower bound}

\item{df_upper}{data.frame: bootstrap upper bound}

\item{ncol_plot}{integer: number of columns in the faceted plot}

\item{dir_plot}{character: path where plot is saved. Obtain using
fm_set_trial and inputting the "dir_plot" object.}

\item{start_k}{numeric: sigmoid steepness (calibrate from 1)}

\item{plot_filename}{character: the name of the plot including the .png
extension.}

\item{start_x0}{numeric: sigmoid midpoint (calibrate from 0.8)}
}
\value{
list: nested list of the identified parameters, errors, thresholds,
and plots.
}
\description{
Visualize the parameter rankings relative to eta star and identify the
type 1 (false negative) and type 2 (false positive) statistical error in
identifying sensitive parameters.
}
\details{
The threshold for parameter identification is determined
by the logistic function, but it generally 0.2 in most cases.
The type 1 error is when the bootstrap mean is above the threshold, and the
lower bound is below the threshold.Type 2 errors are when the bootstrap mean
(mu*) is less than the 0.2 threshold but the upper bound is greater than the
threshold (based on +- standard deviation). The type 1 error is when the
bootstrap mean is above the threshold, and the lower bound is below the
threshold. Generally, a robust sensitivity analysis should be robust to type
2 errors. When the occur, it is recommended to include them as sensitive.
Type 1 errors may also remain included.
}
