% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{fm_project_create}
\alias{fm_project_create}
\title{Create a pwsMultiObsR project}
\usage{
fm_project_create(dir_root, project_name)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}
}
\value{
character: project directory
}
\description{
This function takes your specified working directory and creates a child
directory specific to the project name and relevant input/output directories
}
\examples{
fm_project_create(
  dir_root = "/Users/me/myworkingdir/",
  project_name = "my_project"
)

}
