% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{fm_project_search}
\alias{fm_project_search}
\title{Search pwsMultiObsR projects}
\usage{
fm_project_search(dir_root, verbose = FALSE)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{verbose}{logical: prints the contents of the metadata files}
}
\value{
character vector: list of project names
}
\description{
Print the names of your existing projects
}
\examples{
fm_project_search("/Users/me/myworkingdir/", verbose = TRUE)
}
