% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{fm_trial_create}
\alias{fm_trial_create}
\title{Create a pwsMultiObsR trial directory}
\usage{
fm_trial_create(dir_root, project_name, type, param_names = NULL, nruns = NULL)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{type}{character: string of either "default", "morris", or "mmlhs"}

\item{param_names}{character vector: vector of parameter names to be sampled}

\item{nruns}{integer: number of desired simulations for morris or mmlhs}
}
\value{
list: list of project directories and metadata
}
\description{
Create a trial child directory for a project and generate parameter files.
Note that for large morris or mmlhs trials, it may take a long time for the
function to complete.
}
\examples{
# for a default run
trial_directories <- fm_trial_create(
  dir_root = "/Users/me/myworkingdir/",
  project_name = "my_project",
)

# for morris or mmlhs
trial_directories <- fm_trial_create(
  dir_root = "/Users/me/myworkingdir/",
  project_name = "my_project",
  type = "morris",
  param_names = c("tmax_allsnow", "jh_coef"),
  nruns = 12
)
}
