% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{fm_trial_set}
\alias{fm_trial_set}
\title{Set trial directories}
\usage{
fm_trial_set(dir_root, project_name, trial_number)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{trial_number}{integer: trial number, leading zeros not needed}
}
\value{
list: list of project directories and metadata
}
\description{
Assign a list of trial directories. Used after a trial is created to
continue modeling or processing outputs.
}
\examples{
fm_trial_set(
  dir_root = "/Users/me/myworkingdir/",
  project_name = "my_project",
  trial_number = 2
)

}
