% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_example_inputs}
\alias{load_example_inputs}
\title{Load example pywatershed files and observations}
\usage{
load_example_inputs(dir_root, project_name, example_catchment)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{example_catchment}{character: name of the example catchment, must be
either 'BlueRiv', 'DoloresRiv', 'EastRiv', or 'TaylorRiv'.}
}
\description{
This package contains example model files and observations for the
Blue, Dolores, East, and Taylor river in Colorado, USA. This should be
executed immediately after a project (for one of these catchments) is
created via fm_create_project()
}
