% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_heatmap_eta}
\alias{plot_heatmap_eta}
\title{Plot eta* sensitivity heatmap}
\usage{
plot_heatmap_eta(
  df_eta,
  type1_params = NULL,
  type2_params = NULL,
  param_attributes,
  dir_plot,
  plot_filename = "plot_sensi_heatmap.png"
)
}
\arguments{
\item{df_eta}{data.frame: From the sensi_eet function output, construct
a dataframe of the bootstrap mean eta.star.}

\item{type1_params}{list: (optional) From the fit_logis output, input the
list element named 'type1_params'}

\item{type2_params}{list: (optional) From the fit_logis output, input the
list element named 'type2_params'}

\item{param_attributes}{data.frame: From the fm_set_trial or fm_create_trial
output, use the object named 'param_attributes'}

\item{dir_plot}{character: From the fm_set_trial or fm_create_trial
output, use the object named 'dir_plot'}

\item{plot_filename:}{character: name of the plot with .png extension}
}
\value{
ggplot: plot object
}
\description{
Uses pre-set dataframe to plot a rough heatmap
}
