% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_snotel}
\alias{plot_snotel}
\title{Plot SNOTEL stations in your catchment}
\usage{
plot_snotel(
  dir_root,
  project_name,
  hru_filename = "model_nhru.shp",
  gages_filename = "model_npoigages.shp",
  buffer_dist = 1000,
  map_extent_fac = 0.1,
  plot_filename = "plot_snotel_stations.png",
  option_save = TRUE,
  option_overwrite = FALSE
)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{hru_filename}{character: file name of the hru .shp file}

\item{gages_filename}{character: file name of the gages .shp file}

\item{buffer_dist}{integer: distance in m to flag stations near boundaries}

\item{map_extent_frac}{numeric: fraction of the overall bounding box to
extend the map view. Useful for seeing stations outside of the catchment.}

\item{plot_filename:}{character: name of the plot with .png extension}
}
\value{
ggplot: plot object
}
\description{
Relies on an internal STOTEL station metadata file to find SNOTEL stations
given user specifeid boundaries in a shapefile.
}
\examples{

plot_snotel(
  "/Users/me/mywork/",
  "east_river"
)

}
