% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_usgs}
\alias{plot_usgs}
\title{Plot USGS stream gages in your catchment}
\usage{
plot_usgs(
  dir_root,
  project_name,
  hru_filename = "model_nhru.shp",
  gages_filename = "model_npoigages.shp",
  seg_filename = "model_nsegment.shp",
  map_extent_fac = 0.1,
  plot_filename = "plot_usgs_stations.png",
  option_save = TRUE,
  option_overwrite = FALSE
)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{hru_filename}{character: file name of the hru .shp file}

\item{gages_filename}{character: file name of the gages .shp file}

\item{map_extent_frac}{numeric: fraction of the overall bounding box to
extend the map view. Useful for seeing stations outside of the catchment.}

\item{plot_filename:}{character: name of the plot with .png extension}
}
\value{
ggplot: plot object
}
\description{
Maps USGS stations given user specified PRMS legacy shapefiles.
}
\examples{

plot_usgs("/Users/me/mywork/",
            "east_river")

}
