% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_output}
\alias{read_output}
\title{Read output lists}
\usage{
read_output(
  dir_root,
  project_name,
  trial_number,
  var_names,
  metrics,
  list_type
)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{trial_number}{integer: trial number, leading zeros not needed}

\item{var_names}{character: vector of types e.g. c("q", "swe", "smapsfwt")}

\item{metrics}{character: "gof" of "metric" to specify objective or
predictive functions.}

\item{list_type}{character: "gof" of "eet" to pull objective results or
sensitivity results.}
}
\value{
list: contains corresponding parameters and nested list of outputs.
}
\description{
Recursively compute morris elementary effects and sensitivity indices along
the model outputs (GOF or general statistics)
}
\examples{

# Get data from a single project
gof_data = read_output(
  dir_root = "/Users/me/mywork/",
  project_name = "final_EastRiv",
  trial_number = 3,
  var_names = c("swe","q"),
  metrics = "gof",
  list_type = "gof")

# Can use it in a loop to get data from multiple projects
my_proj_names <- c("final_BlueRiv", "final_DoloresRiv",
                   "final_EastRiv", "final_TaylorRiv")
my_trial_nums <- c(3,3,3,3)
all_output <- list()

for (i in seq_len(my_proj_names)){

  output_name <- paste0("output_",i)

  all_output[output_name] <- read_output(
    dir_root = "/Users/me/mywork/",
    project_name = test_proj_names[i],
    trial_number = test_trial_nums[i],
    var_names = c("swe","q"),
    metrics = "gof",
    list_type = "gof")
}

}
