% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{retrieve_q}
\alias{retrieve_q}
\title{Retrieve USGS daily mean discharge}
\usage{
retrieve_q(
  dir_root,
  project_name,
  usgs_stations,
  start_date,
  end_date,
  overwrite = FALSE
)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{usgs_stations}{list: contains station IDs and respective pws segments}

\item{start_date}{character: in "YYYY-MM-DD" format}

\item{end_date}{character: in "YYYY-MM-DD" format}

\item{overwrite}{logical: TRUE or FALSE to overwrite existing files}
}
\value{
list: retrieved data
}
\description{
Return and save a list of USGS discharge data
}
\examples{

# Input list must follow this format
usgs_stations <- list(site_ids = c("09112500", "09112200"),
                      seg_ids = c(5,4))


start_date <- "1970-10-01"
end_date <- "2023-09-30"

retrieve_q(dir_proj = "/Users/me/mywork/"
           project_name = "East_River",
           usgs_stations = usgs_stations,
           start_date = "1970-10-01",
           end_date = "2023-09-30",
           overwrite = FALSE)

}
