% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pws.R
\name{run_batches}
\alias{run_batches}
\title{Run pywatershed in parallel batches}
\usage{
run_batches(
  dir_root,
  project_name,
  trial_number,
  runs = NULL,
  pwsenvname = "pws",
  out_var_names = c("seg_outflow", "pkwater_equiv", "snowcov_area", "soil_rechr",
    "hru_actet"),
  performance = "medium",
  batch_size = 100,
  pause_time = 60
)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{trial_number}{integer: trial number, leading zeros not needed}

\item{runs}{numeric: a sequence of simulation indices to execute. The
indices correspond to the parameter file indices. E.g. seq(1, nruns, 1).}

\item{pwsenvname}{character: The name of your python environmnent where the
python 'pywatershed' package is installed.}

\item{out_var_names}{character: a vector of the output file names that should
be saved.}

\item{performance}{character:'high', 'medium', or 'low' to select the number
of cores to use. WARNING: this is not intended to be used on high performance
compute clusters or multi-tenant systems.}

\item{batch_size}{integer: the number of of simulations before memory
cleanup.}

\item{pause_time}{integer: time in seconds to shut down the previous
operation and do memory cleanup.}
}
\description{
WARNING: this is not intended to be used on high performance
compute clusters or multi-tenant systems.
}
