% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pws.R
\name{run_default}
\alias{run_default}
\title{Run a default pywatershed simulation}
\usage{
run_default(dir_root, project_name, trial_number, pwsenvname = "pws")
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{trial_number}{integer: trial number, leading zeros not needed}

\item{pwsenvname}{character: The name of your python environmnent where the
python 'pywatershed' package is installed.}

\item{runs}{numeric: a sequence of simulation indices to execute. The
indices correspond to the parameter file indices. E.g. seq(1, nruns, 1).}
}
\description{
WARNING: this is not intended to be used on high performance
compute clusters or multi-tenant systems.
}
