% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morris.R
\name{sensi_eet}
\alias{sensi_eet}
\title{Compute Morris Elementary effect metrics}
\usage{
sensi_eet(
  dir_root,
  project_name,
  trial_number,
  variables = c("swe", "q"),
  metrics = c("gof", "metric"),
  n_boot = 1000,
  sd_level = 1
)
}
\arguments{
\item{dir_root}{character: parent working directory path}

\item{project_name}{character: project name}

\item{trial_number}{integer: trial number, leading zeros not needed}

\item{variables}{character: single or vector of observation types. Must be
one of c("swe", "q", "sms2", "sms8", "sms20", "smapsfwt", "aso", "mod10a1",
"openet")}

\item{metrics}{chracter: single or vector of ouptut list type. Must be one
of c("gof", "metric")}

\item{n_boot}{integer: number of bootstrap replicates of the elementary
effects vector. Set to NA to retrieve baseline results
without bootstrapping (about 100x faster).}

\item{sd_level}{numeric: number of standard devaitions for bootstrap upper
and lower bounds. Should be 1 to match Cuntz et al., 2015, or 1.95 to
match Campolongo et al., 1997. Set to NA to retrieve baseline results
without bootstrapping (about 100x faster).}
}
\value{
list: nested list of metrics for each observation type specified
}
\description{
Recursively compute morris elementary effects and sensitivity indices along
the model outputs (GOF or general statistics)
}
